Added new interface
