+++ src/main/java/core/basesyntax/AreaCalculatable.java
@@ -1,0 +1,5 @@
+package core.basesyntax;
+
+public interface AreaCalculatable {
+    double getArea();
+}
