+++ src/main/java/core/basesyntax/Circle.java
@@ -1,6 +1,6 @@
 package core.basesyntax;
 
-public class Circle extends Shape {
+public class Circle extends Figure {
     private final double radius;
 
     public Circle(String color, int radius) {
