+++ src/main/java/core/basesyntax/Drawable.java
@@ -1,0 +1,5 @@
+package core.basesyntax;
+
+public interface Drawable {
+    void draw();
+}
