+++ src/main/java/core/basesyntax/Figure.java
@@ -1,7 +1,13 @@
 package core.basesyntax;
 
-public interface Figure {
-    double getArea();
+public abstract class Figure implements AreaCalculatable, Drawable {
+    private String color;
 
-    void draw();
+    public Figure(String color) {
+        this.color = color;
+    }
+
+    public String getColor() {
+        return color;
+    }
 }
