+++ src/main/java/core/basesyntax/Main.java
@@ -6,17 +6,17 @@ public class Main {
     public static void main(String[] args) {
         ColorSupplier colorSupplier = new ColorSupplier();
         FigureSupplier figureSupplier = new FigureSupplier(colorSupplier, new Random());
-        Figure[] figures = new Figure[6];
+        Drawable[] figures = new Drawable[6];
 
         for (int i = 0; i < 3; i++) {
             figures[i] = figureSupplier.getFigureRandom();
         }
 
         for (int i = 3; i < 6; i++) {
-            figures[i] = figureSupplier.getDefaultFigure();
+            figures[i] = figureSupplier.getFigureRandom();
         }
 
-        for (Figure figure : figures) {
+        for (Drawable figure : figures) {
             figure.draw();
         }
     }
