+++ src/main/java/core/basesyntax/Rectangle.java
@@ -1,6 +1,6 @@
 package core.basesyntax;
 
-public class Rectangle extends Shape {
+public class Rectangle extends Figure {
     private final double length;
     private final double width;
