+++ src/main/java/core/basesyntax/RightTriangle.java
@@ -1,6 +1,6 @@
 package core.basesyntax;
 
-public class RightTriangle extends Shape {
+public class RightTriangle extends Figure {
     private final double firstLeg;
     private final double secondLeg;
