--- src/main/java/core/basesyntax/Shape.java
@@ -1,15 +1,0 @@
-package core.basesyntax;
-
-public abstract class Shape implements Figure {
-    private String color;
-
-    public Shape(String color) {
-        this.color = color;
-    }
-
-    public String getColor() {
-        return color;
-    }
-
-    public abstract void draw();
-}
